+++ tests/table/apis/table.entity.query.test.ts
@@ -1397,4 +1397,107 @@ describe("table Entity APIs test - using Azure/data-tables", () => {
 
     await tableClient.deleteTable();
   });
+
+  it("23. should find the correct long int, @loki", async () => {
+    const tableClient = createAzureDataTablesClient(
+      testLocalAzuriteInstance,
+      getUniqueName("longint")
+    );
+    const partitionKey = createUniquePartitionKey("");
+    const testEntity: TableTestEntity =
+      entityFactory.createBasicEntityForTest(partitionKey);
+
+    await tableClient.createTable({ requestOptions: { timeout: 60000 } });
+    let result = await tableClient.createEntity(testEntity);
+
+    const anotherPartitionKey = createUniquePartitionKey("");
+    const anotherEntity: TableTestEntity =
+      entityFactory.createBasicEntityForTest(anotherPartitionKey);
+    anotherEntity.int64Field = { value: "1234", type: "Int64" };
+
+    result = await tableClient.createEntity(anotherEntity);
+    assert.ok(result.etag);
+
+    for await (const entity of tableClient
+      .listEntities<TableTestEntity>({
+        queryOptions: {
+          filter: `int64Field gt 1233L and int64Field lt 1235L`
+        }
+      })) {
+      assert.deepStrictEqual(entity.int64Field, 1234n);
+    }
+
+    await tableClient.deleteTable();
+  });
+
+  it("24. should find the correct negative long int, @loki", async () => {
+    const tableClient = createAzureDataTablesClient(
+      testLocalAzuriteInstance,
+      getUniqueName("longint")
+    );
+    const partitionKey = createUniquePartitionKey("");
+    const testEntity: TableTestEntity =
+      entityFactory.createBasicEntityForTest(partitionKey);
+    testEntity.int64Field = { value: "-12345", type: "Int64" };
+
+    await tableClient.createTable({ requestOptions: { timeout: 60000 } });
+    let result = await tableClient.createEntity(testEntity);
+
+    const anotherPartitionKey = createUniquePartitionKey("");
+    const anotherEntity: TableTestEntity =
+      entityFactory.createBasicEntityForTest(anotherPartitionKey);
+    anotherEntity.int64Field = { value: "-1234", type: "Int64" };
+
+    result = await tableClient.createEntity(anotherEntity);
+    assert.ok(result.etag);
+
+    for await (const entity of tableClient
+      .listEntities<TableTestEntity>({
+        queryOptions: {
+          filter: `int64Field lt -1233L and int64Field gt -1235L`
+        }
+      })) {
+      assert.deepStrictEqual(entity.int64Field, -1234n);
+    }
+
+    await tableClient.deleteTable();
+  });
+
+  it("25. should find the correct negative long int, @loki", async () => {
+    const tableClient = createAzureDataTablesClient(
+      testLocalAzuriteInstance,
+      getUniqueName("longint")
+    );
+    const partitionKey = createUniquePartitionKey("");
+    const testEntity: TableTestEntity =
+      entityFactory.createBasicEntityForTest(partitionKey);
+    testEntity.int64Field = { value: "12345", type: "Int64" };
+
+    await tableClient.createTable({ requestOptions: { timeout: 60000 } });
+    let result = await tableClient.createEntity(testEntity);
+
+    const anotherPartitionKey = createUniquePartitionKey("");
+    const anotherEntity: TableTestEntity =
+      entityFactory.createBasicEntityForTest(anotherPartitionKey);
+    anotherEntity.int64Field = { value: "-1234", type: "Int64" };
+
+    result = await tableClient.createEntity(anotherEntity);
+    assert.ok(result.etag);
+
+    let count = 0;
+
+    for await (const entity of tableClient
+      .listEntities<TableTestEntity>({
+        queryOptions: {
+          filter: `int64Field gt -1235L`
+        }
+      })) {
+      entity;
+      ++count;
+    }
+
+    assert.deepStrictEqual(count, 2);
+
+    await tableClient.deleteTable();
+  });
 });